XML configuration of color change
I added XML configuration for setting up the color change updater. XML can be
used to configure either a SingleColorUpdateProvider or a
CountdownColorUpdateProvider.

I also renamed some of the styleable attributes to fit better with the
standard Android attributes naming scheme.

+++ DotMatrixView/src/com/escapeindustries/numericdisplay/CountdownColorUpdateProvider.java
@@ -7,11 +7,11 @@ public class CountdownColorUpdateProvider implements ColorUpdateProvider {
 
 	private int[] lit;
 	private int[] dim;
-	private long[] timeOnColor;
+	private int[] timeOnColor;
 	private long[] changeTimes;
 	private int current;
 
-	public CountdownColorUpdateProvider(int[] lit, int[] dim, long[] timeOnColor) {
+	public CountdownColorUpdateProvider(int[] lit, int[] dim, int[] timeOnColor) {
 		this.lit = lit;
 		this.dim = dim;
 		this.timeOnColor = timeOnColor;
